projects/disentanglement/requirements.txt
torch>=2.1.0+cu123
numpy>=1.22.4
PyYAML>=6.0.1
tqdm>=4.64.0
matplotlib>=3.5.1
seaborn>=0.11.2
imageio>=2.31.5
scikit-learn>=1.3.1
opencv-python>=4.6.0.66
scipy>=1.8.1
six~=1.12.0
Pillow>=9.1.1
torchvision>=0.16.0+cu121
scikit-image>=0.18.3
albumentations>=1.2.0
datasets>=2.15.0
setuptools>=41.2.0
requests>=2.31.0
pandas>=1.4.2
psutil>=5.9.6
tensorboard>=2.4.0
networkx>=3.1
plotly>=5.15.0
dlib>=19.24.0
click>=8.1.3
pyrallis>=0.3.1
torchmetrics>=0.11.4
h5py~=3.10.0
tensorflow>=2.13.0,<2.15.0  # Updated TensorFlow version
keras>=2.13.0,<2.15.0  # Updated Keras version to match TensorFlow
oauth2client~=4.1.3
google-api-python-client~=2.112.0
ninja

